+++ 21-09/app/Spices.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace App;
+
+abstract class Spices
+{
+    public abstract function getName(): string;
+
+}
+
