+++ 21-09/app/Spices/Basil.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Spices;
+
+use App\Spices;
+
+class Basil extends Spices
+{
+
+    public function getName(): string
+    {
+        return 'Basil';
+    }
+}
