+++ 21-09/app/Spices/Oregano.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Spices;
+
+use App\Spices;
+
+class Oregano extends Spices
+{
+
+    public function getName(): string
+    {
+        return 'Oregano';
+    }
+}
