+++ 21-09/app/Spices/Salt.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Spices;
+
+use App\Spices;
+
+class Salt extends Spices
+{
+
+    public function getName(): string
+    {
+        return 'Salt';
+    }
+}
