+++ 21-09/index.php
@@ -1,10 +1,14 @@
 <?php
 
-require_once __DIR__ . '/app/Spices/Spices.php';
+require_once __DIR__ . '/app/Spices.php';
+require_once __DIR__ . '/app/SpicesCollection.php';
+require_once __DIR__ . '/app/Spices/Salt.php';
+require_once __DIR__ . '/app/Spices/Oregano.php';
+require_once __DIR__ . '/app/Spices/Basil.php';
 
 use App\Spices\Salt;
-use App\Spices\Spices;
-use App\Spices\SpicesCollection;
+use App\Spices;
+use App\SpicesCollection;
 use App\Spices\Basil;
 use App\Spices\Oregano;
